feat(cli): implement command normalization and parsing
- Added `normalize.ts` for command definition validation and normalization.
- Introduced `parser.ts` to handle command-line argument parsing and value assignment.
- Created `types.ts` to define types for command options, positional arguments, and parsed values.
- Implemented `usage.ts` for generating human-readable help text for commands.
- Removed the old `index.ts` entry point and adjusted imports across the codebase.
- Updated tests to reflect changes in the CLI parsing and error handling.
- Modified TypeScript configuration to remove deprecated paths and ensure proper module resolution.
- Adjusted build scripts to accommodate new structure and ensure type declarations are correctly generated.
