+++ api/tsconfig.types.json
@@ -2,9 +2,6 @@
     "extends": "./tsconfig.app.json",
     "compilerOptions": {
         "paths": {
-            "catter": [
-                "./build/types/index.d.ts"
-            ],
             "catter/*": [
                 "./build/types/*/index.d.ts",
                 "./build/types/*.d.ts"
